feat: Add classroom images
Add a button to the class component which will show an image of the room

Refs: #15
