+++ timeplan/app/app/globals.module.css
@@ -15,5 +15,5 @@
   right: 0px;
   bottom: 0px;
   height: 20px;
-  width: 150px;
+  width: 70px;
 }
